requirements.txt
streamlit
pandas
numpy
matplotlib
python-dotenv
sqlalchemy
alembic
psycopg2-binary
pymysql
cryptography